source stop semantics
