+++ src/main/java/io/conduit/DestinationService.java
@@ -33,7 +33,7 @@ public class DestinationService extends DestinationPluginGrpc.DestinationPluginI
     private final TaskFactory taskFactory;
     private SinkTask task;
     private Map<String, String> config;
-    private DestinationStream runStream;
+    private DefaultDestinationStream runStream;
     private boolean started;
     private SchemaProvider schemaProvider;
 
@@ -117,7 +117,7 @@ public class DestinationService extends DestinationPluginGrpc.DestinationPluginI
 
     @Override
     public StreamObserver<Destination.Run.Request> run(StreamObserver<Destination.Run.Response> responseObserver) {
-        this.runStream = new DestinationStream(task, schemaProvider, responseObserver);
+        this.runStream = new DefaultDestinationStream(task, schemaProvider, responseObserver);
         return runStream;
     }
